Fix LB malfunction behaviors
The LB should now able to do its work properly, the prerequisite is
that one should configure its machines VIP properly, or the packet
pass through the XDP will be dropped silently, ie. one should add
VIP (virtual IP) manually to the loopback device, and ARP related
settings need to be adjust as well, e.g. :

$ sysctl net.ipv4.conf.all.arp_ignore=1
$ sysctl net.ipv4.conf.eth0.arp_ignore=1
$ sysctl net.ipv4.conf.all.arp_announce=2
$ sysctl net.ipv4.conf.eth0.arp_announce=2

With these adjustment, the LB can do DSR properly with VIP as
source address of the response.

Also, there are bunch of TODOs inside the source, which are all
marked with comment started with `TODO:`.